Add support for entering older dates
If you enter a date in older commits, the entry would always be placed
at the top of the file, Joblog is thus only chronological if every date
you enter is the current day which may not be the case when you're
importing older job applications that were made on much older days.

With this commit, dates you enter will cause Joblog to traverse the
buffer and insert the entry based on its date making it fully
chronological.

Yippe!